requirements.txt
python-socketio==5.9.0
aiohttp==3.8.5
python-dotenv==1.0.0
livekit==1.0.12
livekit-agents==1.2.1
livekit-plugins-cartesia==1.2.1
livekit-plugins-deepgram==1.2.1
livekit-plugins-openai==1.2.1
livekit-plugins-silero==1.2.1
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-turn-detector==1.2.1 